scan the index-resolved positions rarest first, behind a flag
A full-span anchor stops the scan outright, so which position finds one first
decides how much of the rest is done at all. The positions were scanned in
query order; a seed the index put in one place is the likeliest to extend
across the read, and its hit count is already known for the paired stages, so
ordering by it is free there.

It reduces the work: positions scanned fall 4.1% to 119.3M, extensions 4.9%
to 150.6M, with mapped reads unchanged. The time does not follow -- 39.35
against 39.54 worker-seconds is noise -- because the sort costs about what
the skipped extensions save.

Off by default: it changes which anchors survive the per-region cap, so the
output differs and the trade is an accuracy question, not a speed one.

+++ src/minimizer_index.rs
@@ -531,7 +531,7 @@ impl MinimizerIndex {
                 continue;
             };
             for &hit in hits {
-                let ref_id = ((hit >> 48) & 0xffff) as u64;
+                let ref_id = (hit >> 48) & 0xffff;
                 let ref_pos = (hit >> 16) & 0xffff_ffff;
                 packed.push((ref_id << 32) | ref_pos);
             }
